Refactor createNewForm into SubHeader/ utils.ts
